// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as actions_sendDailyDigest from "../actions/sendDailyDigest.js";
import type * as fetchTargets from "../fetchTargets.js";
import type * as http from "../http.js";
import type * as lib from "../lib.js";
import type * as quotes from "../quotes.js";
import type * as routes from "../routes.js";
import type * as syncQuotes from "../syncQuotes.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  "actions/sendDailyDigest": typeof actions_sendDailyDigest;
  fetchTargets: typeof fetchTargets;
  http: typeof http;
  lib: typeof lib;
  quotes: typeof quotes;
  routes: typeof routes;
  syncQuotes: typeof syncQuotes;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;

export declare const components: {};